Closes #16, remove ramsey\uuid dependencies.

+++ app/Containers/Zero/Data/Migrations/2020_07_05_095636_update_ship_job_tables_primary_key_type_to_uuid.php
@@ -11,27 +11,31 @@ class UpdateShipJobTablesPrimaryKeyTypeToUuid extends Migration
      */
     public function up()
     {
-        // Fixes Doctrine Error:
-        // Unknown column type "uuid" requested.
-        DoctrineType::addType('uuid', 'Ramsey\Uuid\Doctrine\UuidType');
+        if(Config::get('queue.default') == 'database'){
 
-        // First drop `id` column, so they can be re-added with new type.
-        Schema::table('jobs', function (Blueprint $table) {
-          $table->dropColumn('id');
-        });
+          // First drop `id` column, so they can be re-added with new type.
+          Schema::table('jobs', function (Blueprint $table) {
+            $table->dropColumn('id');
+          });
 
-        Schema::table('failed_jobs', function (Blueprint $table) {
-          $table->dropColumn('id');
-        });
+          // Now Add `id` columns and update foreign keys
+          Schema::table('jobs', function (Blueprint $table) {
+            $table->uuid('id')->primary();
+          });
+        }
 
-        // Now Add `id` columns and update foreign keys
-        Schema::table('jobs', function (Blueprint $table) {
-          $table->uuid('id')->primary();
-        });
+        if(Config::get('queue.default') == 'database'){
 
-        Schema::table('failed_jobs', function (Blueprint $table) {
-          $table->uuid('id')->primary();
-        });
+          // First drop `id` column, so they can be re-added with new type.
+          Schema::table('failed_jobs', function (Blueprint $table) {
+            $table->dropColumn('id');
+          });
+
+          // Now Add `id` columns and update foreign keys
+          Schema::table('failed_jobs', function (Blueprint $table) {
+            $table->uuid('id')->primary();
+          });
+        }
     }
 
     /**
